Scripting is Key
I created a script (located in /scripts/GenerateJson.py) that will generate all the model files for the strains.

The script is simple, you need to edit the strains array in the script to change the names.

It does not add the strains, just makes it easier to generate new model files.